+++ domain-certificate/main.tf
@@ -1,11 +1,11 @@
 locals {
-  primary_domain_name = var.domain_names[0]
+  primary_domain_name    = var.domain_names[0]
   secondary_domain_names = slice(var.domain_names, 1, length(var.domain_names))
 }
 
 resource "aws_acm_certificate" "default" {
-  domain_name = local.primary_domain_name
-  validation_method = "DNS"
+  domain_name               = local.primary_domain_name
+  validation_method         = "DNS"
   subject_alternative_names = local.secondary_domain_names
 
   tags = merge(var.tags, {
@@ -36,6 +36,6 @@ resource "aws_route53_record" "verification" {
 }
 
 resource "aws_acm_certificate_validation" "default" {
-  certificate_arn = aws_acm_certificate.default.arn
+  certificate_arn         = aws_acm_certificate.default.arn
   validation_record_fqdns = [for record in aws_route53_record.verification : record.fqdn]
 }