+++ ecs_fargate/versions.tf
@@ -1,0 +1,16 @@
+
+terraform {
+  required_version = ">= 0.13"
+
+  required_providers {
+    aws = {
+      source = "hashicorp/aws"
+      # Needed for lambda module
+      version = ">= 4.8.0"
+    }
+    archive = {
+      source  = "hashicorp/archive"
+      version = ">= 2.0"
+    }
+  }
+}