+++ ecscluster/variables.tf
@@ -107,11 +107,17 @@ variable "ami_volumes_delete_on_termination" {
 }
 
 variable "additional_cloudinit_configs" {
-  type = list(string)
+  type        = list(string)
   description = <<-DESC
     Additional cloud-init configuration content to be included in the user_data provided to the cluster's autoscaling group. Additional
     configs are merged with the default cloud-init file using the list(append)+dict(no_replace, recurse_list)+str(append)
     strategy. See https://cloudinit.readthedocs.io/en/latest/reference/merging.html#built-in-mergers for more informaiton.
   DESC
-  default = []
+  default     = []
+}
+
+variable "amazon_ecs_managed_tag" {
+  type        = bool
+  description = "Whether or not to include the AmazonECSManaged tag to the autoscaling group."
+  default     = true
 }