+++ newrelic/alert-conditions-cloudfront/variables.tf
@@ -58,5 +58,5 @@ variable "throughput_threshold" {
 variable "tags" {
   type        = map(any)
   description = "Tags to apply to the alert conditions. Tag values can either be a single string or a list of strings."
-  default = {}
+  default     = {}
 }