+++ newrelic/metric-stream/newrelic.tf
@@ -1,10 +1,3 @@
-resource "newrelic_cloud_aws_link_account" "default" {
-  arn                    = aws_iam_role.newrelic_integration_role.arn
-  metric_collection_mode = "PUSH"
-  name                   = var.newrelic_aws_account_name
-  account_id             = var.newrelic_account_id
-}
-
 resource "aws_iam_role" "newrelic_integration_role" {
   name               = "${var.name_prefix}-newrelic-integration-role"
   assume_role_policy = data.aws_iam_policy_document.newrelic_integration_assume_role.json
@@ -16,7 +9,7 @@ data "aws_iam_policy_document" "newrelic_integration_assume_role" {
     effect = "Allow"
 
     principals {
-      type        = "AWS"
+      type = "AWS"
       # New Relic's AWS account id - from https://docs.newrelic.com/docs/infrastructure/amazon-integrations/connect/connect-aws-new-relic-infrastructure-monitoring/#connect
       identifiers = [754728514883]
     }
@@ -33,7 +26,7 @@ data "aws_iam_policy_document" "newrelic_integration_assume_role" {
 }
 
 resource "aws_iam_role_policy" "newrelic_integration_policy" {
-  role = aws_iam_role.newrelic_integration_role.id
+  role   = aws_iam_role.newrelic_integration_role.id
   policy = data.aws_iam_policy_document.newrelic_integration_policy.json
 }
 
@@ -50,8 +43,8 @@ data "aws_iam_policy_document" "newrelic_integration_policy" {
       "config:BatchGetResourceConfig",
       "config:ListDiscoveredResources",
       "tag:GetResources"
-     ]
+    ]
 
-     resources = ["*"]
+    resources = ["*"]
   }
 }