+++ rdsinstance/main.tf
@@ -1,5 +1,5 @@
-data aws_region default {}
-data aws_caller_identity default {}
+data "aws_region" "default" {}
+data "aws_caller_identity" "default" {}
 
 resource "aws_db_subnet_group" "default" {
   name       = "${var.name}-subnet"
@@ -32,6 +32,10 @@ resource "aws_db_instance" "default" {
   allow_major_version_upgrade           = var.allow_major_version_upgrade
   apply_immediately                     = var.apply_immediately
   iam_database_authentication_enabled   = var.iam_database_authentication_enabled
+  ca_cert_identifier                    = var.ca_cert_identifier
+  manage_master_user_password           = var.manage_master_user_password
+  master_user_secret_kms_key_id         = var.master_user_secret_kms_key_id
+
   vpc_security_group_ids = flatten([
     var.security_groups,
     aws_security_group.db.id,
@@ -124,6 +128,7 @@ data "aws_iam_policy_document" "rds_snapshot_create" {
 }
 
 data "aws_iam_policy_document" "rds_snapshot_delete" {
+  count = var.enable_manual_snapshots ? 1 : 0
   statement {
     effect = "Allow"
     resources = [
@@ -137,7 +142,7 @@ data "aws_iam_policy_document" "rds_snapshot_delete" {
   statement {
     effect = "Allow"
     resources = [
-      "arn:aws:rds:${data.aws_region.default.name}:${data.aws_caller_identity.default.account_id}:snapshot:${aws_db_instance.default.id}*"
+      "arn:aws:rds:${data.aws_region.default.name}:${data.aws_caller_identity.default.account_id}:snapshot:${aws_db_instance.default.identifier}*"
     ]
     actions = [
       "rds:DeleteDBSnapshot"
@@ -148,7 +153,7 @@ data "aws_iam_policy_document" "rds_snapshot_delete" {
 module "backup_lambda" {
   count   = var.enable_manual_snapshots ? 1 : 0
   source  = "github.com/massgov/mds-terraform-common//lambda?ref=1.0.91"
-  name    = "${aws_db_instance.default.id}-backup-lambda"
+  name    = "${aws_db_instance.default.identifier}-backup-lambda"
   package = "${path.module}/dist/backup_lambda.zip"
   handler = "index.handler"
   runtime = "nodejs20.x"
@@ -158,14 +163,14 @@ module "backup_lambda" {
   ]
   environment = {
     variables = {
-      "RDS_INSTANCE_IDENTIFIER" = "${aws_db_instance.default.id}"
+      "RDS_INSTANCE_IDENTIFIER" = "${aws_db_instance.default.identifier}"
     }
   }
   schedule = var.manual_snapshot_schedule
   tags = merge(
     var.tags,
     {
-      "Name" = "${aws_db_instance.default.id}-backup-lambda"
+      "Name" = "${aws_db_instance.default.identifier}-backup-lambda"
     }
   )
   error_topics = var.backup_error_topics
@@ -174,17 +179,17 @@ module "backup_lambda" {
 module "cleanup_lambda" {
   count   = var.enable_manual_snapshots ? 1 : 0
   source  = "github.com/massgov/mds-terraform-common//lambda?ref=1.0.91"
-  name    = "${aws_db_instance.default.id}-cleanup-lambda"
+  name    = "${aws_db_instance.default.identifier}-cleanup-lambda"
   package = "${path.module}/dist/cleanup_lambda.zip"
   handler = "index.handler"
   runtime = "nodejs20.x"
   timeout = 300
   iam_policies = [
-    data.aws_iam_policy_document.rds_snapshot_delete.json
+    data.aws_iam_policy_document.rds_snapshot_delete[count.index].json
   ]
   environment = {
     variables = {
-      "RDS_INSTANCE_IDENTIFIER" = "${aws_db_instance.default.id}"
+      "RDS_INSTANCE_IDENTIFIER" = "${aws_db_instance.default.identifier}"
     }
   }
   schedule = {
@@ -193,7 +198,7 @@ module "cleanup_lambda" {
   tags = merge(
     var.tags,
     {
-      "Name" = "${aws_db_instance.default.id}-cleanup-lambda"
+      "Name" = "${aws_db_instance.default.identifier}-cleanup-lambda"
     }
   )
   error_topics = var.backup_error_topics