+++ vpcread/main.tf
@@ -6,7 +6,7 @@ data "aws_vpc" "default" {
 
 data "aws_subnets" "public" {
   filter {
-    name = "vpc-id"
+    name   = "vpc-id"
     values = [data.aws_vpc.default.id]
   }
   tags = {
@@ -16,7 +16,7 @@ data "aws_subnets" "public" {
 
 data "aws_subnets" "private" {
   filter {
-    name = "vpc-id"
+    name   = "vpc-id"
     values = [data.aws_vpc.default.id]
   }
   tags = {
@@ -26,7 +26,7 @@ data "aws_subnets" "private" {
 
 data "aws_subnets" "db" {
   filter {
-    name = "vpc-id"
+    name   = "vpc-id"
     values = [data.aws_vpc.default.id]
   }
   tags = {